backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.27.2
Pillow>=10.0
anthropic>=0.40.0
google-auth[requests]>=2.35.0
pyjwt>=2.9.0
